Instruct reader to run code before changing code
This lesson is teaching the difference between points versus lines in
ggplot2. The instructions say to first change `geom_point()` to
`geom_line()`. However, if you don't first run the code and plot the
data, you won't be able to see the change between points and lines. It
might be implied to run the code first, but a student may be quick to
start changing things.